requirements.txt
matplotlib~=3.10.0
requests~=2.32.3
numpy~=2.2.2
scipy~=1.15.1
ttkbootstrap~=1.10.1
pillow~=11.1.0